+++ noodles-bam/src/lazy/record/cigar.rs
@@ -1,6 +1,8 @@
 use std::io;
 
-use noodles_sam as sam;
+use noodles_sam::{self as sam, record::cigar::Op};
+
+const CHUNK_SIZE: usize = 4;
 
 /// Raw BAM record CIGAR operations.
 #[derive(Debug, Eq, PartialEq)]
@@ -20,7 +22,20 @@ impl<'a> Cigar<'a> {
     ///
     /// This is _not_ the length of the buffer.
     pub fn len(&self) -> usize {
-        self.0.len() / 4
+        self.0.len() / CHUNK_SIZE
+    }
+
+    /// Returns an iterator over CIGAR operations.
+    pub fn iter(&self) -> impl Iterator<Item = io::Result<Op>> + '_ {
+        use crate::record::codec::decoder::cigar::op::decode_op;
+
+        self.0.chunks(CHUNK_SIZE).map(|chunk| {
+            let buf = chunk
+                .try_into()
+                .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
+            let n = u32::from_le_bytes(buf);
+            decode_op(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+        })
     }
 }
 
@@ -43,5 +58,33 @@ impl<'a> TryFrom<Cigar<'a>> for sam::record::Cigar {
             .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
 
         Ok(cigar)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_iter() -> io::Result<()> {
+        use sam::record::cigar::op::Kind;
+
+        let src = &[][..];
+        let cigar = Cigar::new(src);
+        assert!(cigar.iter().next().is_none());
+
+        let src = &[0x40, 0x00, 0x00, 0x00][..];
+        let cigar = Cigar::new(src);
+        let actual: Vec<_> = cigar.iter().collect::<io::Result<_>>()?;
+        let expected = [Op::new(Kind::Match, 4)];
+        assert_eq!(actual, expected);
+
+        let src = &[0x40, 0x00, 0x00, 0x00, 0x25, 0x00, 0x00, 0x00][..];
+        let cigar = Cigar::new(src);
+        let actual: Vec<_> = cigar.iter().collect::<io::Result<_>>()?;
+        let expected = [Op::new(Kind::Match, 4), Op::new(Kind::HardClip, 2)];
+        assert_eq!(actual, expected);
+
+        Ok(())
     }
 }