+++ noodles-bam/src/record/codec/decoder/cigar.rs
@@ -1,4 +1,4 @@
-mod op;
+pub mod op;
 
 use std::{error, fmt, mem};